Add pages for other types of data than optical and do not require a proposition to be present
