+++ backend/src/GraphQl/Databases/Queries/HasCalorimetricData.graphql
@@ -1,5 +1,5 @@
 query HasCalorimetricData(
-  $where: CalorimetricDataPropositionInput!
+  $where: CalorimetricDataPropositionInput
   $timestamp: DateTime
   $locale: Locale
 ) {
