+++ backend/src/GraphQl/Databases/Queries/HasOpticalData.graphql
@@ -1,5 +1,5 @@
 query HasOpticalData(
-  $where: OpticalDataPropositionInput!
+  $where: OpticalDataPropositionInput
   $timestamp: DateTime
   $locale: Locale
 ) {
