+++ frontend/pages/data/calorimetric.tsx
@@ -1,13 +1,21 @@
-import Layout from "../components/Layout";
-import { Table, message, Form, Button, Alert, Typography } from "antd";
-import { useAllPhotovoltaicDataQuery } from "../queries/data.graphql";
+import Layout from "../../components/Layout";
 import {
-  PhotovoltaicData,
+  Table,
+  message,
+  Form,
+  Button,
+  Alert,
+  Typography,
+  Descriptions,
+} from "antd";
+import { useAllCalorimetricDataQuery } from "../../queries/data.graphql";
+import {
+  CalorimetricData,
   Scalars,
-  PhotovoltaicDataPropositionInput,
-} from "../__generated__/__types__";
+  CalorimetricDataPropositionInput,
+} from "../../__generated__/__types__";
 import { useState } from "react";
-import { setMapValue } from "../lib/freeTextFilter";
+import { setMapValue } from "../../lib/freeTextFilter";
 import {
   getDescriptionColumnProps,
   getFilterableDescriptionListColumnProps,
@@ -15,11 +23,15 @@ import {
   getNameColumnProps,
   getTimestampColumnProps,
   getUuidColumnProps,
-} from "../lib/table";
+} from "../../lib/table";
+import {
+  FloatPropositionComparator,
+  FloatPropositionFormList,
+} from "../../components/FloatPropositionFormList";
 import {
   UuidPropositionComparator,
   UuidPropositionFormList,
-} from "../components/UuidPropositionFormList";
+} from "../../components/UuidPropositionFormList";
 
 // TODO Pagination. See https://www.apollographql.com/docs/react/pagination/core-api/
 
@@ -38,8 +50,8 @@ enum Negator {
 
 const negateIfNecessary = (
   negator: Negator,
-  proposition: PhotovoltaicDataPropositionInput
-): PhotovoltaicDataPropositionInput => {
+  proposition: CalorimetricDataPropositionInput
+): CalorimetricDataPropositionInput => {
   switch (negator) {
     case Negator.Is:
       return proposition;
@@ -50,8 +62,8 @@ const negateIfNecessary = (
 };
 
 const conjunct = (
-  propositions: PhotovoltaicDataPropositionInput[]
-): PhotovoltaicDataPropositionInput => {
+  propositions: CalorimetricDataPropositionInput[]
+): CalorimetricDataPropositionInput => {
   if (propositions.length == 0) {
     return {};
   }
@@ -62,8 +74,8 @@ const conjunct = (
 };
 
 // const disjunct = (
-//   propositions: PhotovoltaicDataPropositionInput[]
-// ): PhotovoltaicDataPropositionInput => {
+//   propositions: CalorimetricDataPropositionInput[]
+// ): CalorimetricDataPropositionInput => {
 //   if (propositions.length == 0) {
 //     return {};
 //   }
@@ -79,12 +91,12 @@ function Page() {
   const [globalErrorMessages, setGlobalErrorMessages] = useState(
     new Array<string>()
   );
-  const [data, setData] = useState<PhotovoltaicData[]>([]);
+  const [data, setData] = useState<CalorimetricData[]>([]);
   // Using `skip` is inspired by https://github.com/apollographql/apollo-client/issues/5268#issuecomment-749501801
   // An alternative would be `useLazy...` as told in https://github.com/apollographql/apollo-client/issues/5268#issuecomment-527727653
   // `useLazy...` does not return a `Promise` though as `use...Query.refetch` does which is used below.
   // For error policies see https://www.apollographql.com/docs/react/v2/data/error-handling/#error-policies
-  const allPhotovoltaicDataQuery = useAllPhotovoltaicDataQuery({
+  const allCalorimetricDataQuery = useAllCalorimetricDataQuery({
     skip: true,
     errorPolicy: "all",
   });
@@ -94,6 +106,8 @@ function Page() {
 
   const onFinish = ({
     componentIds,
+    gValues,
+    uValues,
   }: {
     componentIds:
       | {
@@ -102,12 +116,26 @@ function Page() {
           value: Scalars["UUID"] | undefined;
         }[]
       | undefined;
+    gValues:
+      | {
+          negator: Negator;
+          comparator: FloatPropositionComparator;
+          value: number | undefined;
+        }[]
+      | undefined;
+    uValues:
+      | {
+          negator: Negator;
+          comparator: FloatPropositionComparator;
+          value: number | undefined;
+        }[]
+      | undefined;
   }) => {
     const filter = async () => {
       try {
         setFiltering(true);
         // https://www.apollographql.com/docs/react/networking/authentication/#reset-store-on-logout
-        const propositions: PhotovoltaicDataPropositionInput[] = [];
+        const propositions: CalorimetricDataPropositionInput[] = [];
         if (componentIds) {
           for (let { negator, comparator, value } of componentIds) {
             propositions.push(
@@ -117,7 +145,35 @@ function Page() {
             );
           }
         }
-        const { error, data } = await allPhotovoltaicDataQuery.refetch(
+        // Note that `0` evaluates to `false`, so below we cannot use
+        // `if (value)`.
+        if (gValues) {
+          for (let { negator, comparator, value } of gValues) {
+            if (value !== undefined && value !== null) {
+              propositions.push(
+                negateIfNecessary(negator, {
+                  gValue: {
+                    [comparator]: value,
+                  },
+                })
+              );
+            }
+          }
+        }
+        if (uValues) {
+          for (let { negator, comparator, value } of uValues) {
+            if (value !== undefined && value !== null) {
+              propositions.push(
+                negateIfNecessary(negator, {
+                  uValue: {
+                    [comparator]: value,
+                  },
+                })
+              );
+            }
+          }
+        }
+        const { error, data } = await allCalorimetricDataQuery.refetch(
           propositions.length == 0
             ? {}
             : {
@@ -131,8 +187,13 @@ function Page() {
             error.graphQLErrors.map((error) => error.message).join(" ")
           );
         }
-        // TODO Casting to `PhotovoltaicData` is wrong and error prone!
-        setData((data?.allPhotovoltaicData?.nodes || []) as PhotovoltaicData[]);
+        // TODO Casting to `CalorimetricData` is wrong and error prone!
+        const nestedData =
+          data?.databases?.edges?.map(
+            (edge) => edge?.node?.allCalorimetricData?.nodes || []
+          ) || [];
+        const flatData = ([] as CalorimetricData[]).concat(...nestedData);
+        setData(flatData);
       } catch (error) {
         // TODO Handle properly.
         console.log("Failed:", error);
@@ -149,7 +210,7 @@ function Page() {
 
   return (
     <Layout>
-      <Typography.Title>Photovoltaic Data</Typography.Title>
+      <Typography.Title>Calorimetric Data</Typography.Title>
       {/* TODO Display error messages in a list? */}
       {globalErrorMessages.length > 0 && (
         <Alert type="error" message={globalErrorMessages.join(" ")} />
@@ -162,6 +223,8 @@ function Page() {
         onFinishFailed={onFinishFailed}
       >
         <UuidPropositionFormList name="componentIds" label="Component Id" />
+        <FloatPropositionFormList name="gValues" label="g Values" />
+        <FloatPropositionFormList name="uValues" label="u Values" />
 
         <Form.Item {...tailLayout}>
           <Button type="primary" htmlType="submit" loading={filtering}>
@@ -271,6 +334,20 @@ function Page() {
               ],
               onFilterTextChange,
               (x) => filterText.get(x)
+            ),
+          },
+          {
+            title: "g and u Values",
+            key: "g-and-u-values",
+            render: (_text, record, _index) => (
+              <Descriptions column={1}>
+                <Descriptions.Item key="gValues" label="g Values">
+                  {record.gValues.map((x) => x.toLocaleString("en")).join(", ")}
+                </Descriptions.Item>
+                <Descriptions.Item key="uValues" label="u Values">
+                  {record.uValues.map((x) => x.toLocaleString("en")).join(", ")}
+                </Descriptions.Item>
+              </Descriptions>
             ),
           },
         ]}
